+++ CHANGELOG.md
@@ -34,6 +34,7 @@ Unreleased
 * Removed net adjust_generators.
 * Improved handling of branch and gen outages: Outages can be enabled by setting the outage flag. Contingencies no longer disconnect components.
 * Added bus.is_star(), branch.is_part_of_3_winding_transformer(), and net.get_num_star_buses().
+* Added routines for getting number of gens/branches on outage.
 
 Version 1.3.2
 -------------